Added achievementHandler to methods.js
Added achievements.js and achievementsLogs.js to models.js

Added achievements.js to storage.js

Updated inventory.js (command) to support achievementHandler.js (method)

Updated shop.js (command) to support achievementHandler.js (method)

Updated staff.js (command) to run a check to see if the user is Management (temporarily)

Updated suggestionaccept.js (command) to support achievementHandler.js (method)

Updated xp.js (command) to detect if a user hasn't pinged a user/typed their ID and does nothing instead of throwing an error

Added xpCooldown set in message.js (event) - cooldown can last anywhere betweeen 10 - 20 seconds and prevents earning XP during this time

Added autocreation of achievements and achievementsLogs databases in message.js (event)

Updated message.js (event) to support achievementHandler.js (method)

Added a check in the checkSettings function in message.js (event) to check if the database exists

Fixed multiple issues with floating point numbers ruining the asthetic look of the embeds in message.js (event)

Added achievementHandler.js (method)

Updated startReactionCache.js (method) to [hopefully] fix the issues where roles are re-added upon changing to a different role colour

Added achievements.js (storage)

Added achievementsLogs.js (storage)

Added achievements.js (storage)

Updated inventoryItems.js (storage) with Coin Boosters ranging from +2.5% to +10%

+++ commands/staff.js
@@ -125,6 +125,7 @@ module.exports = class staff extends Multiple(profile, app) {
     }
 
     async run(client, message, args) {
+        if (message.member.roles.get(message.guild.roles.find(x => x.name == "Management").id)) return;
         if (args[1].toLowerCase() == "accept") {
             if (!args[2]) return message.channel.send(`You must specify an ID to accept!`).then((msg) => setTimeout(() => {
                 message.delete();
